pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "overtime"
version = "2.0.0"
description = "Multi-provider homelab infrastructure automation"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "OverTime Contributors"},
]
license = "MIT"

dependencies = [
    "click>=8.1.0",
    "pydantic>=2.5.0",
    "PyYAML>=6.0",
    "paramiko>=3.4.0",
    "python-dotenv>=1.0.0",
    "python-hcl2>=4.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.7.0",
    "black>=23.12.0",
    "ruff>=0.1.8",
]
[project.scripts]
overtime = "overtime.cli:main"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --cov=overtime --cov-report=term-missing"

[tool.setuptools.packages.find]
include = ["overtime*"]
exclude = ["tests*", "docs*", "backups*", "scripts*", "terraform*", "ansible*", "secrets*", "azure*"]
